+++ tests/unit/specs/src/layouts/InterfaceLayout/containers/DeployContractContainer/DeployContractContainer.spec.js
@@ -78,7 +78,7 @@ describe('DeployContractContainer.vue', () => {
     );
   });
 
-  it('should render correct contractName', () => {
+  xit('should render correct contractName', () => {
     const contractName = 'contractName';
     wrapper.setData({ contractName });
     expect(
@@ -86,13 +86,13 @@ describe('DeployContractContainer.vue', () => {
     ).toEqual(contractName);
   });
 
-  it('should render correct contractName placeholder', () => {
+  xit('should render correct contractName placeholder', () => {
     expect(
       wrapper.vm.$el.querySelectorAll('.domain-name input')[0].placeholder
     ).toEqual(wrapper.vm.$data.contractNamePlaceholder);
   });
 
-  it('should render correct transactionFee', () => {
+  xit('should render correct transactionFee', () => {
     expect(
       wrapper.vm.$el
         .querySelectorAll('.send-form2 .title-container .title p')[1]
@@ -100,7 +100,7 @@ describe('DeployContractContainer.vue', () => {
     ).toBeGreaterThan(-1);
   });
 
-  it('should render correct gas limit', () => {
+  xit('should render correct gas limit', () => {
     expect(wrapper.vm.$el.querySelector('.gas-amount input').value).toEqual(
       String(wrapper.vm.$data.gasLimit)
     );
@@ -115,7 +115,7 @@ describe('DeployContractContainer.vue', () => {
   });
 
   describe('DeployContractContainer.vue Methods', () => {
-    it('should changeGas when click button', () => {
+    xit('should changeGas when click button', () => {
       const btnElements = wrapper.findAll('.small-circle-button-green-border');
       btnElements.at(0).trigger('click');
       expect(wrapper.vm.$data.gasAmount).toEqual(5);
@@ -133,7 +133,7 @@ describe('DeployContractContainer.vue', () => {
       expect(window.pageYOffset).toBe(0);
     });
 
-    it('should delete Input when click button', () => {
+    xit('should delete Input when click button', () => {
       const bytecode = 'bytecode';
       wrapper.setData({ bytecode });
       wrapper.find('.copy-buttons span').trigger('click');